cms_form: fix `safe_to_date` to make form extractor happy
Form extractor ignores non required fields if their values is `None`.
In the case of the date field, the util was returning `False`
even if the value was not submitted, leading to an ORM error
whenever the missing field was required.

Now we return `None` and let the extractor deal with proper values
and validation.